task2/stocks_products/requirements.txt
asgiref==3.6.0
django-environ==0.10.0
django-filter==22.1
Django==4.1.7
djangorestframework==3.14.0
environ==1.0
pip==21.3.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.7
wheel==0.37.1
gunicorn